[IOTDB-4826] Move some IT to integration-test (#7891)
