+++ integration-test/src/test/java/org/apache/iotdb/db/it/IoTDBRestartIT.java
@@ -16,91 +16,86 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.integration;
+package org.apache.iotdb.db.it;
 
 import org.apache.iotdb.db.conf.IoTDBConfig;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
-import org.apache.iotdb.db.engine.StorageEngine;
-import org.apache.iotdb.db.exception.StorageEngineException;
-import org.apache.iotdb.db.utils.EnvironmentUtils;
-import org.apache.iotdb.itbase.category.LocalStandaloneTest;
-import org.apache.iotdb.jdbc.Config;
+import org.apache.iotdb.it.env.EnvFactory;
+import org.apache.iotdb.it.framework.IoTDBTestRunner;
+import org.apache.iotdb.itbase.category.ClusterIT;
+import org.apache.iotdb.itbase.category.LocalStandaloneIT;
 
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
 import static org.apache.iotdb.db.constant.TestConstant.TIMESTAMP_STR;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.fail;
 
-@Category({LocalStandaloneTest.class})
+@Ignore
+@RunWith(IoTDBTestRunner.class)
+@Category({LocalStandaloneIT.class, ClusterIT.class})
 public class IoTDBRestartIT {
 
   private final Logger logger = LoggerFactory.getLogger(IoTDBRestartIT.class);
 
   @Before
   public void setUp() throws Exception {
-    EnvironmentUtils.envSetUp();
-    Class.forName(Config.JDBC_DRIVER_NAME);
+    EnvFactory.getEnv().initBeforeClass();
   }
 
   @After
   public void tearDown() throws Exception {
-    EnvironmentUtils.cleanEnv();
+    EnvFactory.getEnv().cleanAfterClass();
   }
 
   @Test
-  public void testRestart() throws SQLException, IOException, StorageEngineException {
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+  public void testRestart() throws SQLException {
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(1,1.0)");
       statement.execute("flush");
     }
 
     try {
-      EnvironmentUtils.restartDaemon();
+      // TODO: replace restartDaemon() with new methods in Env.
+      // EnvironmentUtils.restartDaemon();
     } catch (Exception e) {
-      Assert.fail();
+      fail(e.getMessage());
     }
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(2,1.0)");
     }
 
     try {
-      EnvironmentUtils.restartDaemon();
+      // EnvironmentUtils.restartDaemon();
     } catch (Exception e) {
-      Assert.fail();
+      fail(e.getMessage());
     }
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(3,1.0)");
 
-      boolean hasResultSet = statement.execute("SELECT s1 FROM root.turbine.d1");
-      assertTrue(hasResultSet);
       String[] exp = new String[] {"1,1.0", "2,1.0", "3,1.0"};
       int cnt = 0;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet = statement.executeQuery("SELECT s1 FROM root.turbine.d1")) {
+        assertNotNull(resultSet);
         while (resultSet.next()) {
           String result = resultSet.getString(TIMESTAMP_STR) + "," + resultSet.getString(2);
           assertEquals(exp[cnt], result);
@@ -111,10 +106,8 @@ public class IoTDBRestartIT {
   }
 
   @Test
-  public void testRestartDelete() throws SQLException, IOException, StorageEngineException {
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+  public void testRestartDelete() throws SQLException {
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(1,1)");
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(2,2)");
@@ -122,6 +115,7 @@ public class IoTDBRestartIT {
     }
 
     long time = 0;
+    /*
     try {
       EnvironmentUtils.restartDaemon();
       StorageEngine.getInstance().recover();
@@ -134,19 +128,17 @@ public class IoTDBRestartIT {
         }
       }
     } catch (Exception e) {
-      Assert.fail();
+      fail(e.getMessage());
     }
+     */
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("delete from root.turbine.d1.s1 where time<=1");
 
-      boolean hasResultSet = statement.execute("SELECT s1 FROM root.turbine.d1");
-      assertTrue(hasResultSet);
+      ResultSet resultSet = statement.executeQuery("SELECT s1 FROM root.turbine.d1");
+      assertNotNull(resultSet);
       String[] exp = new String[] {"2,2.0", "3,3.0"};
-      ResultSet resultSet = statement.getResultSet();
       int cnt = 0;
       try {
         while (resultSet.next()) {
@@ -158,10 +150,9 @@ public class IoTDBRestartIT {
         statement.execute("flush");
         statement.execute("delete from root.turbine.d1.s1 where time<=2");
 
-        hasResultSet = statement.execute("SELECT s1 FROM root.turbine.d1");
-        assertTrue(hasResultSet);
         exp = new String[] {"3,3.0"};
-        resultSet = statement.getResultSet();
+        resultSet = statement.executeQuery("SELECT s1 FROM root.turbine.d1");
+        assertNotNull(resultSet);
         cnt = 0;
         while (resultSet.next()) {
           String result = resultSet.getString(TIMESTAMP_STR) + "," + resultSet.getString(2);
@@ -175,49 +166,42 @@ public class IoTDBRestartIT {
   }
 
   @Test
-  public void testRestartQueryLargerThanEndTime()
-      throws SQLException, ClassNotFoundException, IOException, StorageEngineException {
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+  public void testRestartQueryLargerThanEndTime() throws SQLException {
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(1,1)");
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(2,2)");
     }
 
     try {
-      EnvironmentUtils.restartDaemon();
+      // EnvironmentUtils.restartDaemon();
     } catch (Exception e) {
-      Assert.fail();
+      fail(e.getMessage());
     }
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(3,1)");
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(4,2)");
     }
 
     try {
-      EnvironmentUtils.restartDaemon();
+      // EnvironmentUtils.restartDaemon();
     } catch (Exception e) {
-      Assert.fail();
+      fail(e.getMessage());
     }
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet = statement.execute("SELECT s1 FROM root.turbine.d1 where time > 3");
-      assertTrue(hasResultSet);
       String[] exp =
           new String[] {
             "4,2.0",
           };
       int cnt = 0;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet =
+          statement.executeQuery("SELECT s1 FROM root.turbine.d1 where time > 3")) {
+        assertNotNull(resultSet);
         while (resultSet.next()) {
           String result = resultSet.getString(TIMESTAMP_STR) + "," + resultSet.getString(2);
           assertEquals(exp[cnt], result);
@@ -229,46 +213,38 @@ public class IoTDBRestartIT {
   }
 
   @Test
-  public void testRestartEndTime()
-      throws SQLException, ClassNotFoundException, IOException, StorageEngineException {
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+  public void testRestartEndTime() throws SQLException {
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(1,1)");
       statement.execute("insert into root.turbine.d1(timestamp,s1) values(2,2)");
     }
 
     try {
-      EnvironmentUtils.restartDaemon();
+      // EnvironmentUtils.restartDaemon();
     } catch (Exception e) {
-      Assert.fail();
+      fail(e.getMessage());
     }
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine.d1(timestamp,s2) values(1,1)");
       statement.execute("insert into root.turbine.d1(timestamp,s2) values(2,2)");
     }
 
     try {
-      EnvironmentUtils.restartDaemon();
+      // EnvironmentUtils.restartDaemon();
     } catch (Exception e) {
-      Assert.fail();
+      fail(e.getMessage());
     }
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet = statement.execute("SELECT s2 FROM root.turbine.d1");
-      assertTrue(hasResultSet);
       String[] exp = new String[] {"1,1.0", "2,2.0"};
       int cnt = 0;
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet = statement.executeQuery("SELECT s2 FROM root.turbine.d1")) {
+        assertNotNull(resultSet);
         while (resultSet.next()) {
           String result = resultSet.getString(TIMESTAMP_STR) + "," + resultSet.getString(2);
           assertEquals(exp[cnt], result);
@@ -281,9 +257,7 @@ public class IoTDBRestartIT {
 
   @Test
   public void testRecoverWALMismatchDataType() throws Exception {
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine1.d1(timestamp,s1,s2) values(1,1.1,2.2)");
       statement.execute("delete timeseries root.turbine1.d1.s1");
@@ -291,22 +265,19 @@ public class IoTDBRestartIT {
           "create timeseries root.turbine1.d1.s1 with datatype=INT32, encoding=RLE, compression=SNAPPY");
     }
 
-    EnvironmentUtils.restartDaemon();
+    // EnvironmentUtils.restartDaemon();
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet = statement.execute("select * from root.**");
-      assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet = statement.executeQuery("select * from root.**")) {
+        assertNotNull(resultSet);
         int cnt = 0;
-        Assert.assertEquals(3, resultSet.getMetaData().getColumnCount());
+        assertEquals(3, resultSet.getMetaData().getColumnCount());
         while (resultSet.next()) {
-          Assert.assertEquals("1", resultSet.getString(1));
-          Assert.assertEquals(null, resultSet.getString(2));
-          Assert.assertEquals("2.2", resultSet.getString(3));
+          assertEquals("1", resultSet.getString(1));
+          assertNull(resultSet.getString(2));
+          assertEquals("2.2", resultSet.getString(3));
           cnt++;
         }
         assertEquals(1, cnt);
@@ -316,29 +287,24 @@ public class IoTDBRestartIT {
 
   @Test
   public void testRecoverWALDeleteSchema() throws Exception {
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine1.d1(timestamp,s1,s2) values(1,1.1,2.2)");
       statement.execute("delete timeseries root.turbine1.d1.s1");
     }
 
-    EnvironmentUtils.restartDaemon();
+    // EnvironmentUtils.restartDaemon();
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet = statement.execute("select * from root.**");
-      assertTrue(hasResultSet);
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet = statement.executeQuery("select * from root.**")) {
+        assertNotNull(resultSet);
         int cnt = 0;
-        Assert.assertEquals(2, resultSet.getMetaData().getColumnCount());
+        assertEquals(2, resultSet.getMetaData().getColumnCount());
         while (resultSet.next()) {
-          Assert.assertEquals("1", resultSet.getString(1));
-          Assert.assertEquals("2.2", resultSet.getString(2));
+          assertEquals("1", resultSet.getString(1));
+          assertEquals("2.2", resultSet.getString(2));
           cnt++;
         }
         assertEquals(1, cnt);
@@ -356,9 +322,7 @@ public class IoTDBRestartIT {
     config.setSeqTsFileSize(10000000);
     config.setUnSeqTsFileSize(10000000);
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("create timeseries root.turbine1.d1.s1 with datatype=INT64");
       statement.execute("insert into root.turbine1.d1(timestamp,s1) values(1,1)");
@@ -369,16 +333,15 @@ public class IoTDBRestartIT {
     }
 
     Thread.sleep(1000);
-    EnvironmentUtils.restartDaemon();
+    // EnvironmentUtils.restartDaemon();
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
       long[] result = new long[] {1L, 2L};
-      statement.execute("select s1 from root.turbine1.d1 where time < 3");
-      ResultSet resultSet = statement.getResultSet();
+      ResultSet resultSet =
+          statement.executeQuery("select s1 from root.turbine1.d1 where time < 3");
+      assertNotNull(resultSet);
       int cnt = 0;
       while (resultSet.next()) {
         assertEquals(resultSet.getLong(1), result[cnt]);
@@ -394,33 +357,27 @@ public class IoTDBRestartIT {
 
   @Test
   public void testRecoverFromFlushMemTableError() throws Exception {
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
       statement.execute("insert into root.turbine1.d1(timestamp,s1,s2) values(1,1.1,2.2)");
     }
 
     // mock exception during flush memtable
-    EnvironmentUtils.restartDaemon();
+    // EnvironmentUtils.restartDaemon();
 
-    try (Connection connection =
-            DriverManager.getConnection(
-                Config.IOTDB_URL_PREFIX + "127.0.0.1:6667/", "root", "root");
+    try (Connection connection = EnvFactory.getEnv().getConnection();
         Statement statement = connection.createStatement()) {
 
-      boolean hasResultSet = statement.execute("select * from root.**");
-      assertTrue(hasResultSet);
-
-      try (ResultSet resultSet = statement.getResultSet()) {
+      try (ResultSet resultSet = statement.executeQuery("select * from root.**")) {
+        assertNotNull(resultSet);
         int cnt = 0;
         while (resultSet.next()) {
-          Assert.assertEquals("1", resultSet.getString(1));
-          Assert.assertEquals("1.1", resultSet.getString(2));
-          Assert.assertEquals("2.2", resultSet.getString(3));
+          assertEquals("1", resultSet.getString(1));
+          assertEquals("1.1", resultSet.getString(2));
+          assertEquals("2.2", resultSet.getString(3));
           cnt++;
         }
-        Assert.assertEquals(1, cnt);
+        assertEquals(1, cnt);
       }
     }
   }
