+++ integration-test/src/test/java/org/apache/iotdb/session/it/pool/SessionPoolIT.java
@@ -18,8 +18,8 @@
  */
 package org.apache.iotdb.session.it.pool;
 
-import org.apache.iotdb.commons.conf.IoTDBConstant;
 import org.apache.iotdb.db.utils.EnvironmentUtils;
+import org.apache.iotdb.it.env.EnvFactory;
 import org.apache.iotdb.it.framework.IoTDBTestRunner;
 import org.apache.iotdb.rpc.IoTDBConnectionException;
 import org.apache.iotdb.rpc.StatementExecutionException;
@@ -30,8 +30,8 @@ import org.apache.iotdb.session.util.Version;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
@@ -60,18 +60,19 @@ public class SessionPoolIT {
 
   @Before
   public void setUp() throws Exception {
-    System.setProperty(IoTDBConstant.IOTDB_CONF, "src/test/resources/");
+    EnvFactory.getEnv().initBeforeTest();
     EnvironmentUtils.envSetUp();
   }
 
   @After
   public void tearDown() throws Exception {
     EnvironmentUtils.cleanEnv();
+    EnvFactory.getEnv().cleanAfterClass();
   }
 
   @Test
   public void insert() {
-    SessionPool pool = new SessionPool("127.0.0.1", 6667, "root", "root", 3);
+    SessionPool pool = EnvFactory.getEnv().getSessionPool(3);
     ExecutorService service = Executors.newFixedThreadPool(10);
     for (int i = 0; i < 10; i++) {
       final int no = i;
@@ -104,7 +105,7 @@ public class SessionPoolIT {
 
   @Test
   public void incorrectSQL() {
-    SessionPool pool = new SessionPool("127.0.0.1", 6667, "root", "root", 3);
+    SessionPool pool = EnvFactory.getEnv().getSessionPool(3);
     assertEquals(0, pool.currentAvailableSize());
     try {
       pool.insertRecord(
@@ -123,7 +124,7 @@ public class SessionPoolIT {
 
   @Test
   public void incorrectExecuteQueryStatement() {
-    SessionPool pool = new SessionPool("127.0.0.1", 6667, "root", "root", 3);
+    SessionPool pool = EnvFactory.getEnv().getSessionPool(3);
     ExecutorService service = Executors.newFixedThreadPool(10);
     write10Data(pool, true);
     // now let's query
@@ -134,7 +135,7 @@ public class SessionPoolIT {
             try {
               SessionDataSetWrapper wrapper =
                   pool.executeQueryStatement("select * from root.sg1.d1 where time = " + no);
-              // this is incorrect becasue wrapper is not closed.
+              // this is incorrect because wrapper is not closed.
               // so all other 7 queries will be blocked
             } catch (IoTDBConnectionException | StatementExecutionException e) {
               fail(e.getMessage());
@@ -156,14 +157,14 @@ public class SessionPoolIT {
 
   @Test
   public void executeQueryStatement() {
-    SessionPool pool = new SessionPool("127.0.0.1", 6667, "root", "root", 3);
+    SessionPool pool = EnvFactory.getEnv().getSessionPool(3);
     correctQuery(pool, DEFAULT_QUERY_TIMEOUT);
     pool.close();
   }
 
   @Test
   public void executeQueryStatementWithTimeout() {
-    SessionPool pool = new SessionPool("127.0.0.1", 6667, "root", "root", 3);
+    SessionPool pool = EnvFactory.getEnv().getSessionPool(3);
     correctQuery(pool, 2000);
     pool.close();
   }
@@ -206,7 +207,7 @@ public class SessionPoolIT {
 
   @Test
   public void executeRawDataQuery() {
-    SessionPool pool = new SessionPool("127.0.0.1", 6667, "root", "root", 3);
+    SessionPool pool = EnvFactory.getEnv().getSessionPool(3);
     ExecutorService service = Executors.newFixedThreadPool(10);
     write10Data(pool, true);
     List<String> pathList = new ArrayList<>();
@@ -218,7 +219,7 @@ public class SessionPoolIT {
             try {
               SessionDataSetWrapper wrapper = pool.executeRawDataQuery(pathList, no, no + 1, 60000);
               if (wrapper.hasNext()) {
-                Assert.assertEquals(no, wrapper.next().getTimestamp());
+                assertEquals(no, wrapper.next().getTimestamp());
               }
               pool.closeResultSet(wrapper);
             } catch (Exception e) {
@@ -240,6 +241,7 @@ public class SessionPoolIT {
   }
 
   @Test
+  @Ignore
   public void tryIfTheServerIsRestart() {
     SessionPool pool =
         new SessionPool(
@@ -261,6 +263,7 @@ public class SessionPoolIT {
     SessionDataSetWrapper wrapper = null;
     try {
       wrapper = pool.executeQueryStatement("select * from root.sg1.d1 where time > 1");
+      // TODO: replace stopDaemon() and restartDaemon() with new methods in Env.
       EnvironmentUtils.stopDaemon();
       // user does not know what happens.
       while (wrapper.hasNext()) {
@@ -339,6 +342,7 @@ public class SessionPoolIT {
   }
 
   @Test
+  @Ignore
   public void tryIfTheServerIsRestartButDataIsGotten() {
     SessionPool pool =
         new SessionPool(
@@ -379,6 +383,7 @@ public class SessionPoolIT {
   }
 
   @Test
+  @Ignore
   public void restart() {
     SessionPool pool =
         new SessionPool(
@@ -444,22 +449,7 @@ public class SessionPoolIT {
 
   @Test
   public void testClose() {
-    SessionPool pool =
-        new SessionPool(
-            "127.0.0.1",
-            6667,
-            "root",
-            "root",
-            3,
-            1,
-            60000,
-            false,
-            null,
-            false,
-            SessionConfig.DEFAULT_CONNECTION_TIMEOUT_MS,
-            SessionConfig.DEFAULT_VERSION,
-            SessionConfig.DEFAULT_INITIAL_BUFFER_CAPACITY,
-            SessionConfig.DEFAULT_MAX_FRAME_SIZE);
+    SessionPool pool = EnvFactory.getEnv().getSessionPool(3);
     pool.close();
     try {
       pool.insertRecord(
@@ -469,7 +459,7 @@ public class SessionPoolIT {
           Collections.singletonList(TSDataType.INT64),
           Collections.singletonList(1L));
     } catch (IoTDBConnectionException e) {
-      Assert.assertEquals("Session pool is closed", e.getMessage());
+      assertEquals("Session pool is closed", e.getMessage());
     } catch (StatementExecutionException e) {
       fail(e.getMessage());
     }
